removed SparseState, DenseState
