+++ rusty_lr_core/src/builder/state.rs
@@ -48,7 +48,7 @@ where
     Term: Ord,
 {
     fn from(state: crate::builder::State<Term, NonTerm>) -> Self {
-        use crate::parser::state::ShiftTarget;
+        use crate::parser::table::ShiftTarget;
 
         crate::parser::state::IntermediateState {
             shift_goto_map_term: state
